requirements.txt
Flask==2.2.3
folium==0.14.0
geopy==2.3.0
python-dotenv==1.0.0
requests==2.28.1
